requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
fastapi==0.78.0
h11==0.12.0
httpcore==0.15.0
idna==3.3
loguru==0.6.0
pydantic==1.9.1
pymongo==4.1.1
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
win32-setctime==1.1.0
